% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{addGenes}
\alias{addGenes}
\title{Add columns with distance to nearest gene and gene symbol(s)}
\usage{
addGenes(query, geneset, genome, cachedir)
}
\arguments{
\item{query}{Genomic regions to find nearest genes for as a GRanges, data.frame, or data.table with the columns "chr", "start", and "end"}

\item{geneset}{Select one of "ucsc" for the UCSC Genes (from the knownGene table), "refseq" for RefSeq genes (from the refFlat table), or "ensembl" for the Ensembl genes (from the ensGene table)}

\item{genome}{UCSC genome name to use (e.g. hg19, mm10)}

\item{cachedir}{Path where cached UCSC tables are stores}
}
\description{
Add columns with distance to nearest gene and gene symbol(s)
}

