% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.r
\name{addNearest}
\alias{addNearest}
\title{Add columns to query with distance to nearest subject and subject id(s)}
\usage{
addNearest(query, subject, id = "name", prefix = "subject")
}
\arguments{
\item{query}{Genomic regions to find nearest genes for as a GRanges, data.frame, or data.table with the columns "chr", "start", and "end"}

\item{id}{Column name of the id field in subject to report as the nearest id(s). In case of ties, a comma separated list will be returned.}

\item{prefix}{Append this string to names of the added columns}

\item{query}{Genomic regions to find nearest genes for as a GRanges, data.frame, or data.table with the columns "chr", "start", and "end"}
}
\description{
Add columns to query with distance to nearest subject and subject id(s)
}

