% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichment.r
\name{calcEnrichmentBinom}
\alias{calcEnrichmentBinom}
\title{Perform binomial test of enrichment for motifs using counts of occurrences in two sequence sets}
\usage{
calcEnrichmentBinom(seq1.counts, seq1.nSeqs, seq2.counts, seq2.nSeqs)
}
\arguments{
\item{seq1.counts}{output object (matrix) from \code{\link{calcMotifCounts}} for first sequence set}

\item{seq1.nSeqs}{number of sequences in first sequence set}

\item{seq2.counts}{output object from \code{\link{calcMotifCounts}} for second sequence set}

\item{seq2.nSeqs}{number of sequences in second sequence set}
}
\value{
dataframe of output results including p-values (unadjusted)
}
\description{
The *.counts matrix objects must first be generated using \code{\link{calcMotifCounts}}. The current implementation only considers if a sequence has at least one occurrence of the motif or not, and does not account for or weight multiple occurrences of a motif in a single sequence. The contingency table is simply based on the number of sequences which contain at least one occurrence of each motif.
}

