% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffmotif.R
\name{calcMotifCounts}
\alias{calcMotifCounts}
\title{Generate counts of motif occurrences in each sequence}
\usage{
calcMotifCounts(fimo.out, q.cutoff)
}
\arguments{
\item{fimo.out}{dataframe from \code{\link{readFIMO}} dataframe object}

\item{q.cutoff}{only count a motif if the q-value is less than this cutoff}
}
\value{
matrix with pairwise counts of each motif and each sequence
}
\description{
Generate counts of motif occurrences in each sequence
}

