% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichment.r
\name{doPropMatch}
\alias{doPropMatch}
\title{Do PSM}
\usage{
doPropMatch(target, pool, pdf = NULL, formula, n = 1, bsg, genome, cachedir)
}
\arguments{
\item{obj}{A data.frame or data.table with columns "chr", "start", and "end" and any other columns}
}
\value{
A GRanges made from the data in obj.
}
\description{
PSM function
}

