% Generated by roxygen2 (4.0.1): do not edit by hand
\name{drawBackgroundSetPropensity}
\alias{drawBackgroundSetPropensity}
\title{Draw a matched reference set from a reference pool}
\usage{
drawBackgroundSetPropensity(target.seq, target.meta, pool.seq, pool.meta,
  formula, start.order)
}
\arguments{
\item{target.seq}{\code{DNAStringSet} object of the target set}

\item{target.meta}{\code{data.frame} object with the covariates of the target set}

\item{pool.seq}{\code{DNAStringSet} object of the reference pool to draw covariate matched reference set from}

\item{pool.meta}{\code{data.frame} object with the covariates}

\item{formula}{an \code{as.formula} object for the regression used to generate propensity scores}

\item{start.order}{vector of starting order for matching (must be a sequence of integers in any order from 1 to the total number of sequences in both target.seq and pool.seq)}
}
\value{
\code{DNAStringSet} object of a covariate-matched reference set
}
\description{
Use propensity score matching to create a covariate-matched reference set. Note that the matching function is sensitive to the starting order of the input data. This order is required as a variable so it can be fixed between runs.
}

