% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.r
\name{getCpgFeatures}
\alias{getCpgFeatures}
\title{Generate feature sets based on CpG island, shore, and shelf regions}
\usage{
getCpgFeatures(genome, cachedir)
}
\arguments{
\item{genome}{The UCSC name specific to the genome of the query coordinates (e.g. "hg19", "hg18", "mm10", etc)}

\item{cachedir}{A path to a directory where a local cache of UCSC tables are stored. If equal to \code{NULL} (default), the data will be downloaded to temporary files and loaded on the fly. Caching is highly recommended to save time and bandwidth.}
}
\description{
Uses the "cpgIslandExt" table to generate shore (+/- 2kb from islands) and shelf (+/- 2kb from shores) regions. Will only function for genomes with this table available. Can be concatenated using c() with other tables, such as from getFeatures(), and provided as input for the "features" argument of goldmine().
}

