% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFeatures}
\alias{getFeatures}
\title{Obtain feature sets from UCSC genome browser tables}
\usage{
getFeatures(tables = c("wgEncodeRegDnaseClusteredV2",
  "wgEncodeRegTfbsClusteredV3", "tfbsConsSites", "cosmic", "oreganno",
  "vistaEnhancers", "phastConsElements100way"), genome, cachedir)
}
\arguments{
\item{tables}{A vector of table names from UCSC (default: set of useful tables).}

\item{genome}{See goldmine()}

\item{cachedir}{See goldmine()}
}
\description{
Given a vector of table names from the UCSC genome browser that all contain "chrom", "chromStart", and "chromEnd" fields, converts them to input suitable for the goldmine() "features" argument.
}

