% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.r
\name{getGeneModels}
\alias{getGeneModels}
\title{Return sets of ranges for individual gene model components based on the contents of the output from getGenes()}
\usage{
getGeneModels(genes = getGenes(geneset = "ucsc", genome = genome, cachedir =
  cachedir), genome, cachedir)
}
\arguments{
\item{genes}{Genes of interest from the output table of getGenes(). If not given, will default to the UCSC knownGene table.}

\item{genome}{The UCSC name specific to the genome of the query coordinates (e.g. "hg19", "hg18", "mm10", etc)}

\item{cachedir}{A path to a directory where a local cache of UCSC tables are stored. If equal to \code{NULL} (default), the data will be downloaded to temporary files and loaded on the fly. Caching is highly recommended to save time and bandwidth.}
}
\value{
A list containing one GenomicRanges object for each of the gene model portions: Promoters, 3' Ends, Exons, Introns, Intergenic, 5' UTRs, 3' UTRs. The "srow" column can be used to match individual ranges to individual genes in the table given to the "genes" argument by row number.
}
\description{
For custom analysis that requires the genomic ranges of gene model components.
}

