% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffmotif.R
\name{getSeqMeta}
\alias{getSeqMeta}
\title{Calculate covariates for each sequence in a DNAStringSet}
\usage{
getSeqMeta(ranges, bsgenome, genome, cachedir)
}
\arguments{
\item{myseq}{DNAStringSet object of the sequence set}
}
\value{
dataframe with standard covariates added
}
\description{
Calculate covariates for each sequence in a DNAStringSet
}

