% Generated by roxygen2 (4.0.2): do not edit by hand
\name{makeDT}
\alias{makeDT}
\title{Make a data.table from a GRanges or a data.frame}
\usage{
makeDT(obj)
}
\arguments{
\item{obj}{A data.frame or GRanges}
}
\value{
A data.table made from the data in obj.
}
\description{
Given a data.frame or GRanges, a data.table object will be created. If the input is already a data.table, it is simply returned.
}

