% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffmotif.R
\name{plotCovarDistance}
\alias{plotCovarDistance}
\title{Plot horizontal graph of covariate distance from different propensity models}
\usage{
plotCovarDistance(orig.meta, list.meta, cols)
}
\arguments{
\item{orig.meta}{dataframe of covariates from the target set you are trying to match}

\item{list.meta}{a list of dataframes of covariates from other sets you want to compare to the target set}

\item{cols}{vector of which columns to use from the dataframes above}
}
\value{
plots to active graphics device
}
\description{
Horizontal graph plots a point for each variable that represents the distance between that variable's value in orig.meta and each of the dataframes in list.meta
}

