% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffmotif.R
\name{plotCovarHistogramsOverlap}
\alias{plotCovarHistogramsOverlap}
\title{Plot overlapping histograms for any number of variables from 2 sets}
\usage{
plotCovarHistogramsOverlap(seq1.meta, seq2.meta, cols, plot.ncols = 3,
  main = "")
}
\arguments{
\item{seq1.meta}{dataframe of covariates from first distribution}

\item{seq2.meta}{dataframe of covariates from second distribution}

\item{cols}{which columns in the *.meta dataframes contain covariate data to plot}

\item{plot.ncols}{number of columns in the plotted grid}

\item{main}{title for the grid of plots (useful if you want to put on the formula you used to generate seq2.meta)}
}
\value{
plot to active graphics device
}
\description{
Plots a grid of overlapping histograms. Data for seq1 will appear in red and seq2 in blue. The region where the distributions will appear in purple.
}

