% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffmotif.R
\name{plotCovarQQ}
\alias{plotCovarQQ}
\title{Plot QQ plots in a grid for arbitrary number of variables}
\usage{
plotCovarQQ(orig.meta, list.meta, cols, plot.ncols = 3)
}
\arguments{
\item{orig.meta}{data frame from the original distribution}

\item{list.meta}{list of data frames for all the distributions to compare to for each variable}

\item{cols}{which columns have covariates to plot from the above dataframes}

\item{plot.ncols}{how many columns the plotted grid should have}
}
\value{
plot to active graphics device
}
\description{
Creates QQ plots to compare two distributions for any number of variables
}

