% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffmotif.R
\name{readFIMO}
\alias{readFIMO}
\title{Read in a FIMO output file}
\usage{
readFIMO(fimo.out.path)
}
\arguments{
\item{fimo.out.path}{path to FIMO output text file}
}
\description{
Wrapper of read.table() with correct options for reading in a FIMO output text file.
}

