% Generated by roxygen2 (4.0.2): do not edit by hand
\name{readUCSCAnnotation}
\alias{readUCSCAnnotation}
\title{Read UCSC table files from disk and join all related tables}
\usage{
readUCSCAnnotation(genome = "hg19", path = "")
}
\description{
Read UCSC table files from disk and join all related tables
}

