% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffmotif.R
\name{runFIMO}
\alias{runFIMO}
\title{Wrapper to call FIMO using system()}
\usage{
runFIMO(out.path, fasta.path, motifs.path)
}
\arguments{
\item{out.path}{path where output will be written}

\item{fasta.path}{path to FASTA file of input sequences}

\item{motifs.path}{path to MEME format file containing motif database to use}
}
\description{
If FIMO from the MEME Suite is installed and in the current PATH, this provides an easy interface
to run it from R.
}

