% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.r
\name{sortGRanges}
\alias{sortGRanges}
\title{Sort a data.frame, data.table, or GRanges by chr (accounting for mixed string and numeric names), start, end and return a GRanges}
\usage{
sortGRanges(obj)
}
\arguments{
\item{obj}{A data.frame, data.table, or GRanges}
}
\description{
Sort a data.frame, data.table, or GRanges by chr (accounting for mixed string and numeric names), start, end and return a GRanges
}

