% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/enrichment.r
\name{drawGenomePool}
\alias{drawGenomePool}
\title{Draw a length-matched pool of sequences from the genome}
\usage{
drawGenomePool(query, n, genome, cachedir, chrs = NULL)
}
\value{
A GRanges of the background sequences.
}
\description{
Given a query set of ranges, draw a length-matched pool of sequences.
}

