% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getCpgFeatures}
\alias{getCpgFeatures}
\title{Generate feature sets based on CpG island, shore, and shelf regions}
\usage{
getCpgFeatures(genome, cachedir)
}
\arguments{
\item{genome}{See goldmine()}

\item{cachedir}{See goldmine()}
}
\description{
Uses the "cpgIslandExt" table to generate shore (+/- 2kb from islands) and shelf (+/- 2kb from shores) regions.
}

