% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add.r
\name{getGenes}
\alias{getGenes}
\title{Load table of gene ranges via UCSC Genome Browser tables}
\usage{
getGenes(geneset = "ucsc", genome, cachedir = NULL)
}
\arguments{
\item{geneset}{Select one of "ucsc" for the UCSC Genes (from the knownGene table), "refseq" for RefSeq genes (from the refFlat table), or "ensembl" for the Ensembl genes (from the ensGene table)}

\item{genome}{UCSC genome name to use (e.g. hg19, mm10)}

\item{cachedir}{Path where cached UCSC tables are stores}
}
\description{
Load table of gene ranges via UCSC Genome Browser tables
}

