% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{gmWrite}
\alias{gmWrite}
\title{Write individual CSV files to disk from the output of goldmine()}
\usage{
gmWrite(gm, path = ".")
}
\arguments{
\item{gm}{The output list object from goldmine().}

\item{path}{The directory to write the files into (default: current working directory).}
}
\description{
Write a CSV file for each output table in a goldmine() output list object.
}

