% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldmine-package.r
\docType{package}
\name{goldmine-package}
\alias{goldmine-package}
\title{goldmine: Genomic context annotation for any set of genomic ranges using UCSC Genome Browser tables}
\description{
goldmine: Genomic context annotation for any set of genomic ranges using UCSC Genome Browser tables
}

