% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotCovarHistograms}
\alias{plotCovarHistograms}
\title{Plot histograms in a grid for arbitrary number of variables}
\usage{
plotCovarHistograms(seq.meta, cols)
}
\arguments{
\item{seq.meta}{data.frame of sequence covariates}

\item{cols}{vector of which columns to plot histograms for from seq.meta}
}
\value{
plot sent to current graphics device
}
\description{
Plots non-overlapping single histograms in a grid for all covariate data in a dataframe.
}

