% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ucsc.r
\name{readUCSCAnnotation}
\alias{readUCSCAnnotation}
\title{Read UCSC table files from disk and join all related tables}
\usage{
readUCSCAnnotation(genome = "hg19", path = "")
}
\description{
Read UCSC table files from disk and join all related tables
}

