% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sortDT}
\alias{sortDT}
\title{Sort a data.frame, data.table, or GRanges by chr (accounting for mixed string and numeric names), start, end and return a data.table}
\usage{
sortDT(obj)
}
\arguments{
\item{obj}{A data.frame, data.table, or GRanges}
}
\description{
Sort a data.frame, data.table, or GRanges by chr (accounting for mixed string and numeric names), start, end and return a data.table
}

