% Generated by roxygen2 (4.0.2): do not edit by hand
\name{testEnrichment}
\alias{testEnrichment}
\title{Make a GRanges from a data.frame or data.table with the fields "chr", "start", and "end"}
\usage{
testEnrichment(query, background, features)
}
\arguments{
\item{obj}{A data.frame or data.table with columns "chr", "start", and "end" and any other columns}
}
\value{
A GRanges made from the data in obj.
}
\description{
Given a data.frame or data.table with the columns "chr", "start", and "end", a GenomicRanges (GRanges) object will be created. All other columns will be passed on as metadata. If the input is already a GRanges, it is simply returned.
}

