% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{writeBEDFromGRanges}
\alias{writeBEDFromGRanges}
\title{Write a BED format file from a GenomicRanges object}
\usage{
writeBEDFromGRanges(gr, file, name = NULL)
}
\arguments{
\item{gr}{A GenomicRanges object.}

\item{file}{Filename of the BED file to write.}

\item{name}{Column name to use for the name field in the BED file (optional)}
}
\description{
Creates BED file suitable for upload as a custom track to the UCSC genome browser. Note that start coordinates are 0-based in the BED format.
}

